+++ connector_test.go
@@ -1,6 +1,3 @@
-//go:build go1.10
-// +build go1.10
-
 package proxy
 
 import (
