+++ find_caller.go
@@ -1,5 +1,4 @@
 //go:build go1.12
-// +build go1.12
 
 package proxy
